+++ pymongo/mongo_client.py
@@ -411,7 +411,9 @@ class MongoClient(common.BaseObject):
           - `authMechanismProperties`: Used to specify authentication mechanism
             specific options. To specify the service name for GSSAPI
             authentication pass authMechanismProperties='SERVICE_NAME:<service
-            name>'
+            name>'.
+            To specify the session token for MONGODB-AWS authentication pass
+            ``authMechanismProperties='AWS_SESSION_TOKEN:<session token>'``.
 
           .. seealso:: :doc:`/examples/authentication`
 
@@ -520,7 +522,7 @@ class MongoClient(common.BaseObject):
 
         .. versionchanged:: 3.5
            Add ``username`` and ``password`` options. Document the
-           ``authSource``, ``authMechanism``, and ``authMechanismProperties ``
+           ``authSource``, ``authMechanism``, and ``authMechanismProperties``
            options.
            Deprecated the ``socketKeepAlive`` keyword argument and URI option.
            ``socketKeepAlive`` now defaults to ``True``.